+++ docs/README.md
@@ -48,8 +48,8 @@
 - [x] alert는 void형이기 때문에, 삼항연산자 지양 
 - [x] hidden없애고 id로 하기 
 - [x] 상수 (하드코딩 x)
+- [x] alert파일 데이터 분리 
 - [ ] 파일 정리해보기 (여러개인것 합치기)
-- [ ] 만약 재시작한다면 init? (드림코딩 참고)
 
 ## ✅ 요구 사항
 - [x] 주어진 index.html에 html 엘리먼트를 직접 추가하거나 기존의 html 엘리먼트를 임의로 삭제하지 않는다. id와 같은 선택자를 추가하는 작업만 가능하다.
